написал сервисы без маппера

+++ src/main/java/ru/skypro/homework/dto/CommentDto.java
@@ -15,7 +15,7 @@ public class CommentDto {
     private String authorFirstName;
 
     @Schema(type = "integer", format = "int64", description = "дата и время создания комментария в миллисекундах с 00:00:00 01.01.1970")
-    private Integer createdAt;
+    private Long createdAt;
 
     @Schema(type = "integer", format = "int32", description = "id комментария")
     private Integer pk;
@@ -26,7 +26,7 @@ public class CommentDto {
     public CommentDto(Integer author,
                       String authorImage,
                       String authorFirstName,
-                      Integer createdAt,
+                      Long createdAt,
                       Integer pk,
                       String text) {
         this.author = author;
@@ -64,11 +64,11 @@ public class CommentDto {
         this.authorFirstName = authorFirstName;
     }
 
-    public Integer getCreatedAt() {
+    public Long getCreatedAt() {
         return createdAt;
     }
 
-    public void setCreatedAt(Integer createdAt) {
+    public void setCreatedAt(Long createdAt) {
         this.createdAt = createdAt;
     }
